+++ src/main/java/com/unascribed/yttr/content/enchant/DisjunctionEnchantment.java
@@ -1,12 +1,17 @@
 package com.unascribed.yttr.content.enchant;
 
+import org.jetbrains.annotations.Nullable;
+
 import com.unascribed.yttr.YConfig;
+import com.unascribed.yttr.init.YTags;
+import com.unascribed.yttr.mixinsupport.ComplexDamageEnchant;
 
 import net.minecraft.enchantment.DamageEnchantment;
+import net.minecraft.entity.Entity;
 import net.minecraft.entity.EntityGroup;
 import net.minecraft.entity.EquipmentSlot;
 
-public class DisjunctionEnchantment extends DamageEnchantment {
+public class DisjunctionEnchantment extends DamageEnchantment implements ComplexDamageEnchant {
 
 	public static final EntityGroup ENDER = new EntityGroup();
 	
@@ -27,10 +32,15 @@ public class DisjunctionEnchantment extends DamageEnchantment {
 	}
 	
 	@Override
-	public float getAttackDamage(int level, EntityGroup group) {
-		if (group == ENDER) {
-			return level * 3.5f;
+	public AttackResult handleAttack(int level, Entity target, @Nullable Entity attacker) {
+		if (target.getType().isIn(YTags.Entity.DISJUNCTIBLE)) {
+			return new AttackResult(null, level * 3.5f);
 		}
+		return null;
+	}
+	
+	@Override
+	public float getAttackDamage(int level, EntityGroup group) {
 		return 0;
 	}