+++ src/main/java/com/unascribed/yttr/mixinsupport/ClassNodeTransformer.java
@@ -1,0 +1,9 @@
+package com.unascribed.yttr.mixinsupport;
+
+import org.objectweb.asm.tree.ClassNode;
+
+public interface ClassNodeTransformer {
+
+	void transform(String name, ClassNode node);
+	
+}